+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
  * Fixed default size for Object and Active Object on sequence diagram 
  * Fixed In Scope on sequence diagram
  * Fixed UML Primitive behavior when resizing
+ * Added Terminator element to Sequence Diagram.
 
 ## [1.1.2-beta.3] - 2025-03-10
 ### Fixed